Backend/pyproject.toml
[project]
name = "linkedin-comments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.0.0",
    "uvicorn>=0.35.0",
]
